backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1
elasticsearch==8.11.0
kafka-python==2.0.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
psycopg2-binary==2.9.9
geoalchemy2==0.14.2
shapely==2.0.2
geopy==2.4.0
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Authentication & Security
slowapi==0.1.9
cryptography==41.0.7
authlib==1.2.1
email-validator>=1.1.0,<2.1

# Rate Limiting & CORS  
fastapi-limiter==0.1.6